More setup. RPC testing. I now understand how to use RPCs for features. Widget testing for multiplayer.
RPC requires the use of an event or a function that is first called and has the replication set to server and then this calls the main event or function with the actual logic which has replication for multi.
For this project I will be designating server calls with the identifier (server) and multi replication with the identifier (multi) so that they can be easily defined and called.

Created a test widget to test my understanding of component replication for the health bar of the players.

Can now start actually working on systems and core gameplay just also need to figure out how to do actual multiplayer eventually when not in editor.