Fixed editable text in create session widget. Organisation. Rebased game mode and game state to their non-base versions so can use multiplayer specific functions.
https://www.youtube.com/watch?v=Hsr6mbNKBLU&list=PLNb7FZ2Nw2HTcJ9Qvy8n2Ou-ZVbsDOMFh&index=5
Really good video for understanding the different classes for multiplayer.
Including: game state, player state, pawns, controllers, game modes and widgets for multiplayer.
Reclassed some of the classes already created to be the multiplayer versions instead of the base versions.

Spawning players from event rather than automatically.
It's like I begin to understand all this and then something happens that crumbles everything and I get confused again.

Spawning players is as follows:
Debug button on player controller - must be local player controller and must be on server so the host which then calls spawn all players in the game mode.
Then in the game mode we loop through the player controllers and call spawn player in the controller and then update spawned info in the controller.
Player controller spawn player first gets game mode and calls spawn player and then sets input mode game and UI and then sets show mouse cursor and enable click events true. Update spawned info calls it for multi which then gets the player state and casts and then calls update player info in the player state.
Update player info in the player state validates the passed controller and then calls setup player in the controlled pawn and then casts to the player to then call update camera for player in the player state. This then updates the view target with blend and updates the camera ref as well as setting look at cursor true after 1 second.

New events in the player for setting up the player using server and multi depending on has authority.

There must be an easier way to do this.
May update it eventually but this seems to be working for now.
Would obviously need to advance this more with more menus for changing appearance and then also the host starting the game via clickable button rather than debug key.