Lobby. Need more edge case functionality.
https://www.youtube.com/watch?v=PcfJiD6RTmo&list=PLNb7FZ2Nw2HTcJ9Qvy8n2Ou-ZVbsDOMFh&index=9
Used this video to help understand widgets for multiplayer.

New player controller, game mode and widgets for the lobby.
Currently the lobby just shows the connected player's names and pings and has a leave session button.

Things that still need implementing:
Loading screens.
Host leaving lobby / session then properly resetting all clients so that they go back to the main menu.
Can't leave whilst still connecting or whilst a client is still connecting if we are the host of the game.