Fixed resolution problem. Widget tweaking appearance. YAY host leaving fix. Auto updating lobby info periodically so always up to date as a backup.
There was an issue with the widget for the main menu not showing properly on other clients but changing resolution of the windows changed this for some reason.
Removed backup calling for main menu widget creation and addition to screen instead just using PC_Lobby to do this so not duplicating menu widgets.

Game instance printing errors using built in functions.
Turns out the host disconnecting all players is just built into the game if you set the game startup map to be the one you actually want (main menu level) since disconnecting goes from like almost startup again I think.
Using event tick of the lobby game mode to periodically update the lobby info of all stored player controllers. This is only done every 0.25 seconds and its in the lobby so not really that big of a deal I dont think.

Tomorrow will work on the aesthetics, readying up, starting the game from the lobby, changing player info in the lobby, loading screens and other aesthetics.