More widgets. Removed delay for showing lobby info. Loading screen. Difference for lobby info for host and self.
Error message widget and functionality setup.
Created in instance when error occurs. Passes error enum to change text.
Close button which then sets the game mode back. Canvas panel is visible so cannot interact with anything behind it.

Removed the delay in the PC_Lobby for updating lobby info since doing it repeatedly in the game mode. Event tick more frequent in the lobby game mode.

Loading screen widget created and functionality for showing it when entering the lobby using the player controller for lobby.
Updated functionality for updating lobby info whilst loading screen is active lobby widget is not shown.

Lobby info now changes depending on if the info is the host or self.
Changing colour and also having a (host) or (self) before the name of the person in the lobby.