Player Lobby Customisation.
Players now also have colourful representations of themselves in the lobby.
Input actions and functionality implemented for readying up and changing colours cycling through available colours without having repeating colours. The only edge case can't check for is if two players change colours at the same time but that shouldn't really cause game breaking bugs and would just result in them having the same colour which could hopefull be fixed by them then changing colours.

New widgets for the player customisation.
One for the individual and another containing a wrap box which then has the individuals as children within it.
Lots of functionality in the individual customisation widget, lobby player controller and game mode for this to work as intended.
Lots of trial and error but eventually got it working.

Now want the individual customisations names to have the self and host tags where necessary.