Player customisation name (self) and (host) tags. Find sessions auto updating info that is already displayed.
In player customisation now also uses similar functionality as player ping / name for showing the self and host tags in the customisation parts of the lobby.

Find sessions now auto updates.
There is one issue with this, when a host closes their session then it automatically takes on the information of another session without allowing joining however this is not really an issue since you just refresh the searches and it is removed so yea.
Tried for quite a while to validate the session data so could fix this but couldn't get it to work properly.