Loading into game (sorta fix). Waiting for players screen. Updating camera properly. Can't load in if only one player in lobby. Map Introduction widget and setting controls. Game flow / progression continuing.
After all controllers loaded into the game we have a 2 second delay currently to make sure that everything is loaded properly and ready.
Start game function in the game mode spawns the players with corresponding info since the player info is now a exposed on spawn variable that then updates the info on begin play for the player.
Then after looping through spawning each character we then update the camera for each controller and then call the fade waiting screen function in each controller.

New widgets in this push: waiting screen, map intro.
Waiting screen is active whilst waiting for controllers to load and the game to start.
Map intro is at the start of each map like in STRAFTAT.
Widget animations for fading the waiting screen and one for the map intro.

New events and functionality in the player controller and active game mode for the map and activating and deactivating input.
Whilst map intro active no input other than mouse movement is active.

When fading the waiting screen we get the camera ref from the player state and then update it in the player.
Will need to do functionality eventually for updating the camera ref for the specific map in the player state and then also updating it every time new map is moved to for the player controller and character.

Now moving onto more functionality for the game flow like rounds ending, counting scores, and moving to new maps.