Upgrades Animation Syncing. Round and Map Win widget and syncing. Fixed issue with resetting players. Round / map win indicators synced and updating properly with round and map wins.
Upgrade uncover, hover, unhover and select animations syncing across all players.
Also disabling upgrade widgets for players that aren't the ones picking so they definitely can't pick.
Uncover animation is cool.
Setting the text on the upgrade after the uncover animation is done.

New widgets for displaying the round and map wins, all in one display like in rounds with half and full circles in the player colours.
New functionality in the game mode, player controller and the widgets themselves for all this.

Instead of healing the players when they reset we now just set the current health to maximum health since healing is disabled whilst below or at 0 health so it wasn't actually healing.