Actually showing player name in the upgrade screen. Upgrade randomisation actually working and synced properly now as well.
Using player name properly instead of using controller name for the upgrade screen.

New enum entries in upgrade names for placeholders.
New functionality in the game mode for active game for calculating the upgrades to be able to be chosen from.
As long as there are more than 5 upgrades then there won't be any problems with duplicates.
Unique upgrades for each one and synced across all player clients as well.