Weapon attachment update. Was trying to change melee combat method but just going to use different animations. Attack cooldown. Need juice. Attack Hitboxes for weapons.
Weapon attaching for holstering and for attacking.
Updating melee component for attack cooldown.
Need to juice the melee combat.
Attacking hitboxes setup and in use for melee combat now with damage and knockback for anything with health component.

Noticed that need to setup functionality for starting the hazards at the same time since the saw blades are out of sync if on begin play.