Camera replicating. Projectile pooling in ranged component. Replicated projectile pooling.
Camera start and stop events are replicating multi. Also walls reset is multi replicating as well.
Functionality in the ranged component for spawning a set amount of projectiles.
Don't need to replicate the actual spawning of the projectiles since the projectiles themselves replicate.
Currently calling the start projectile pooling in begin play of the component. May continue doing this or have the game mode call it IDK.
The array of projectiles stored in the ranged component is replicated so that they have the same references across server and client.
New interface for projectile.
Activate and deactivate events in ranged component and projectile actor.
Event in ranged component for updating projectile info.
New struct for projectile info currently only containing a float of speed.
Need to get the projectile actually moving and the rotation correct but is mostly working and replicating properly.